enhancing the deleteJob function

+++ server/src/modules/jobs/jobs.controller.js
@@ -97,8 +97,9 @@ const updateJob = async (req, res) => {
             return res.status(403).send("Only the owner can edit this Job!");
         }
 
-        if (foundJob.status !== 'open' ||foundJob.status !== 'draft') {
-            return res.status(404).json({ err: `Cannot edit this job, due to it being ${foundJob.status}` })        }
+        if (foundJob.status !== 'open' || foundJob.status !== 'draft') {
+            return res.status(400).json({ err: `Cannot edit this job, due to it being ${foundJob.status}` })
+        }
 
         const updatedJob = await Job.findByIdAndUpdate(req.params.jobId, req.body, { returnDocument: 'after' })
 
@@ -124,6 +125,10 @@ const deleteJob = async (req, res) => {
 
         if (!foundJob.client.equals(req.user._id)) {
             return res.status(403).send("Only the owner can edit this Job!");
+        }
+        
+        if (foundJob.status !== 'draft') {
+            return res.status(400).json({ err: `Cannot edit this job, due to it being ${foundJob.status}` })
         }
 
         const deletedJob = await Job.findByIdAndDelete(req.params.jobId)
